Use Self Signed Cert-Friendy SSL Context in Integration Test
